StayController : getStaysMadeByUser, Exceptions

+++ src/main/java/finalproject/airbnb/controller/StayController.java
@@ -1,6 +1,8 @@
 package finalproject.airbnb.controller;
 
 
+import finalproject.airbnb.exceptions.AuthorizationException;
+import finalproject.airbnb.exceptions.NotFoundException;
 import finalproject.airbnb.model.dao.StayDAO;
 import finalproject.airbnb.model.dto.GetStayDTO;
 import finalproject.airbnb.model.dto.StayDTO;
@@ -11,17 +13,20 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
 import java.sql.SQLException;
+import java.util.List;
 
 @RestController
-public class StayController {
+public class StayController extends AbstractController {
 
     @Autowired
     private StayDAO stayDAO;
 
     @PostMapping("/stays")
     public Stay addStay(@RequestBody StayDTO stayDTO , HttpSession session) throws SQLException {
-        //User user = (User) session.getAttribute(UserController.SESSION_KEY_LOGGED_USER);
-        User user = null;
+        User user = (User) session.getAttribute(UserController.SESSION_KEY_LOGGED_USER);
+        if(user == null) {
+            throw new AuthorizationException();
+        }
         Stay stay = new Stay(stayDTO);
         stay.setHost(user);
         stayDAO.addStay(stay);
@@ -32,17 +37,38 @@ public class StayController {
     public GetStayDTO getStay(@PathVariable long id) throws SQLException {
         GetStayDTO getStayDTO = stayDAO.getStayById(id);
         if(getStayDTO == null) {
-            //no stay
+            throw new NotFoundException("Stay not found");
         }
         return getStayDTO;
     }
 
     @DeleteMapping("/stays/{id}")
-    public String deleteStay(@PathVariable long id) throws SQLException {
+    public String deleteStay(@PathVariable long id, HttpSession session) throws SQLException {
+        User user = (User) session.getAttribute(UserController.SESSION_KEY_LOGGED_USER);
+        if(user == null) {
+            throw new AuthorizationException();
+        }
+        if(user.getId() != stayDAO.getHostId(id)) {
+            throw new AuthorizationException("You don't have permissions to delete this stay!");
+        }
         if(stayDAO.getStayById(id) == null) {
-            //no stay
+            throw new NotFoundException("Stay not found");
         }
         return stayDAO.deleteStay(id);
     }
+
+    @GetMapping("/users/{id}/stays")
+    public List<GetStayDTO> getStaysMadeByUser(@PathVariable long id, HttpSession session) throws SQLException {
+        User user = (User) session.getAttribute(UserController.SESSION_KEY_LOGGED_USER);
+        if(user == null) {
+            throw new AuthorizationException();
+        }
+        List<GetStayDTO> stays = stayDAO.getStaysByUserId(id);
+        if(stays.isEmpty()) {
+            throw new NotFoundException("Stays not found");
+        }
+        return stays;
+    }
+
 
 }
